rename: 상품리스트 관련 UI 패키지 경로 수정
